src/requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.2
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.13.1
opencv_python==4.7.0.68
Pillow==9.4.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
psycopg2-binary
Werkzeug==2.2.2
pytest
pytest-flask
locust==2.14.2
